requirements.txt
requests
python-dotenv
streamlit
flask
pandas
numpy
altair
pydeck
pillow
protobuf
typing-extensions
langchain_core
langchain_community
openai
simple-salesforce
pytest
langchain_community
langchain_openai
langchain_salesforce
langchain 

